+++ src/file-lint/file-lint.class.ts
@@ -3,7 +3,7 @@ import { readFile } from 'fs';
 import { IOptions, sync } from 'glob';
 import { resolve } from 'path';
 import { isArray, promisify } from 'util';
-import { IZConfigReader } from '../lint-config/config-reader.interface';
+import { IZConfigReader } from '../config/config-reader.interface';
 import { IZLinter } from '../lint/linter.interface';
 import { IZContentLinter } from './content-linter.interface';